reorganize component files
