+++ app/components/library/DexIndicator.jsx
