+++ app/components/pages/Profile/index.jsx
@@ -4,20 +4,20 @@ import { useDispatch, useSelector } from 'react-redux';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router';
 
-import { DexCreate } from './dex-create';
-import { DexPreview } from './dex-preview';
-import { Footer } from './footer';
-import { FriendCode } from './friend-code';
-import { Header } from './header';
-import { Nav } from './nav';
-import { NotFound } from './not-found';
-import { Notification } from './notification';
-import { Reload } from './reload';
-import { checkVersion } from '../actions/utils';
-import { listGames } from '../actions/game';
-import { retrieveUser, setCurrentUser, setUser } from '../actions/user';
-import { setShowShare } from '../actions/tracker';
-import { listDexTypes } from '../actions/dex-type';
+import { DexCreate } from './DexCreate';
+import { DexPreview } from './DexPreview';
+import { Footer } from '../../library/Footer';
+import { FriendCode } from '../../library/FriendCode';
+import { Header } from '../../library/Header';
+import { Nav } from '../../library/Nav';
+import { NotFound } from '../NotFound';
+import { Notification } from '../../library/Notification';
+import { Reload } from '../../library/Reload';
+import { checkVersion } from '../../../actions/utils';
+import { listGames } from '../../../actions/game';
+import { retrieveUser, setCurrentUser, setUser } from '../../../actions/user';
+import { setShowShare } from '../../../actions/tracker';
+import { listDexTypes } from '../../../actions/dex-type';
 
 export function Profile () {
   const dispatch = useDispatch();
