+++ app/components/pages/Tracker/InfoLocations.jsx
