+++ app/index.tsx
@@ -5,7 +5,7 @@ import { ErrorBoundary, Provider as RollbarProvider } from '@rollbar/react';
 import { Provider } from 'react-redux';
 import { render } from 'react-dom';
 
-import { App } from './components/app';
+import { App } from './components/pages/App';
 import { Rollbar } from './utils/rollbar';
 import { Store } from './stores';
